refactor(copilot): extract CopilotResult and collect_copilot_response to sdk/collect.py
Move CopilotResult and collect_copilot_response out of the 1772-line
service.py into a dedicated collect.py module. These are a public API
for consuming copilot streams and deserve their own file. Update
__init__.py lazy imports and autopilot.py import accordingly.
