+++ autogpt_platform/backend/backend/copilot/sdk/__init__.py
@@ -24,8 +24,8 @@ from typing import TYPE_CHECKING, Any
 # Static imports for type checkers so they can resolve __all__ entries
 # without executing the lazy-import machinery at runtime.
 if TYPE_CHECKING:
-    from .service import CopilotResult as CopilotResult
-    from .service import collect_copilot_response as collect_copilot_response
+    from .collect import CopilotResult as CopilotResult
+    from .collect import collect_copilot_response as collect_copilot_response
     from .service import stream_chat_completion_sdk as stream_chat_completion_sdk
     from .tool_adapter import create_copilot_mcp_server as create_copilot_mcp_server
 
@@ -39,8 +39,8 @@ __all__ = [
 # Dispatch table for PEP 562 lazy imports.  Each entry is a (module, attr)
 # pair so new exports can be added without touching __getattr__ itself.
 _LAZY_IMPORTS: dict[str, tuple[str, str]] = {
-    "CopilotResult": (".service", "CopilotResult"),
-    "collect_copilot_response": (".service", "collect_copilot_response"),
+    "CopilotResult": (".collect", "CopilotResult"),
+    "collect_copilot_response": (".collect", "collect_copilot_response"),
     "stream_chat_completion_sdk": (".service", "stream_chat_completion_sdk"),
     "create_copilot_mcp_server": (".tool_adapter", "create_copilot_mcp_server"),
 }
